ci: add dockerfile with nginx server
I created an nginx dockerfile and set the expose keyword to 80 to notify
the runner that 80 is expected. I also added a simple index.html file
that says hello-world simply to test the output and ensure the website
is being exposed to the host.

The nginx config is basic and one I just got from a course I used to
come up with this process

This can be tested using
docker run --publish 80:80 website