ci: add docker compose script
I added this in order to facilitate basic orchestration via docker
compose. This will do the work being done before by running the docker
command. The docker command will still be executed but instead of docker
run, the command is now:

docker compose up website

This is basically a wrapper that can manage more than one docker
container to create an orchestration, or can just be used to manage the
lifespan of any number of containers, even a single one.